requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
google-generativeai==0.3.2
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0